* Changed the type of the logicStateChanged signal (an enum may be the same   size as an int).
* Replaced an occurrence of LinkManager::instance where it should be
  LinkManager::hasInstance().


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/branches/logiclinking@2392 f8843c6b-282e-0410-804b-c0d5895947d2

+++ src/svxlink/svxlink/LinkManager.h
@@ -145,7 +145,7 @@ class LinkManager : public sigc::trackable
                             const std::string &subcmd);
     std::vector<std::string> getCommands(std::string logicname) const;
 
-    sigc::signal<void, uint8_t, bool> logicStateChanged;
+    sigc::signal<void, int, bool> logicStateChanged;
 
   private:
     class CGuard
